in middle of getting routes/controllers/server.js to work for registerUser.js
